services/ai-inference/requirements.txt
ultralytics>=8.0.0
onnxruntime-gpu>=1.16.0
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0

fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

grpcio>=1.60.0
grpcio-tools>=1.60.0
protobuf>=4.25.0

prometheus-client>=0.19.0
structlog>=23.2.0

norfair>=2.2.0
shapely>=2.0.0

PyYAML>=6.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
